requirements.txt
pandas
matplotlib
seaborn
plotly
streamlit
python-dotenv
groq
openpyxl